Move SSORevokeButton to components folder
